BREAKING(io): move types file
