+++ io/buf_writer.ts
@@ -2,7 +2,7 @@
 // This module is browser compatible.
 
 import { copy } from "../bytes/copy.ts";
-import type { Writer, WriterSync } from "./types.d.ts";
+import type { Writer, WriterSync } from "./types.ts";
 
 const DEFAULT_BUF_SIZE = 4096;
