+++ io/buffer.ts
@@ -3,7 +3,7 @@
 
 import { assert } from "../assert/assert.ts";
 import { copy } from "../bytes/copy.ts";
-import type { Reader, ReaderSync } from "./types.d.ts";
+import type { Reader, ReaderSync } from "./types.ts";
 
 // MIN_READ is the minimum ArrayBuffer size passed to a read call by
 // buffer.ReadFrom. As long as the Buffer has at least MIN_READ bytes beyond
