+++ io/limited_reader.ts
@@ -7,7 +7,7 @@
  * `read` returns `null` when `limit` <= `0` or
  * when the underlying `reader` returns `null`.
  */
-import type { Reader } from "./types.d.ts";
+import type { Reader } from "./types.ts";
 
 /**
  * @deprecated (will be removed after 1.0.0) Use the [Web Streams API]{@link https://developer.mozilla.org/en-US/docs/Web/API/Streams_API} instead.
