+++ io/read_string_delim.ts
@@ -1,7 +1,7 @@
 // Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
 // This module is browser compatible.
 
-import { type Reader } from "./types.d.ts";
+import { type Reader } from "./types.ts";
 import { readDelim } from "./read_delim.ts";
 
 /**
