+++ io/string_writer.ts
@@ -1,7 +1,7 @@
 // Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
 // This module is browser compatible.
 
-import type { Writer, WriterSync } from "./types.d.ts";
+import type { Writer, WriterSync } from "./types.ts";
 
 const decoder = new TextDecoder();
