+++ streams/readable_stream_from_reader_test.ts
@@ -5,7 +5,7 @@ import { readableStreamFromReader } from "./readable_stream_from_reader.ts";
 import { Buffer } from "../io/buffer.ts";
 import { concat } from "../bytes/concat.ts";
 import { copy } from "../bytes/copy.ts";
-import type { Closer, Reader } from "../io/types.d.ts";
+import type { Closer, Reader } from "../io/types.ts";
 
 class MockReaderCloser implements Reader, Closer {
   chunks: Uint8Array[] = [];
