+++ streams/write_all.ts
@@ -1,7 +1,7 @@
 // Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
 // This module is browser compatible.
 
-import type { Writer, WriterSync } from "../io/types.d.ts";
+import type { Writer, WriterSync } from "../io/types.ts";
 export type { Writer, WriterSync };
 
 /**
